remove new Client class, use Recognizer
