+++ src/common.speech/SpeakerServiceRecognizer.ts
@@ -194,8 +194,8 @@ export class SpeakerServiceRecognizer extends ServiceRecognizerBase {
 
         const sessionStartEventArgs: SessionEventArgs = new SessionEventArgs(this.privRequestSession.sessionId);
 
-        if (!!this.privClient.sessionStarted) {
-            this.privClient.sessionStarted(this.privClient, sessionStartEventArgs);
+        if (!!this.privRecognizer.sessionStarted) {
+            this.privRecognizer.sessionStarted(this.privRecognizer, sessionStartEventArgs);
         }
 
         void this.receiveMessage();
