+++ src/sdk/IntentRecognizer.ts
@@ -10,6 +10,7 @@ import {
     RecognitionMode,
     RecognizerConfig,
     ServiceRecognizerBase,
+    SpeechServiceConfig
 } from "../common.speech/Exports";
 import { marshalPromiseToCallbacks } from "../common/Exports";
 import { AudioConfigImpl } from "./Audio/AudioConfig";
@@ -304,6 +305,10 @@ export class IntentRecognizer extends Recognizer {
         Contracts.throwIfDisposed(this.privDisposedIntentRecognizer);
 
         marshalPromiseToCallbacks(this.dispose(true), cb, errorCb);
+    }
+
+    protected createRecognizerConfig(speechConfig: SpeechServiceConfig): RecognizerConfig {
+        return new RecognizerConfig(speechConfig, this.privProperties);
     }
 
     protected createServiceRecognizer(authentication: IAuthentication, connectionFactory: IConnectionFactory, audioConfig: AudioConfig, recognizerConfig: RecognizerConfig): ServiceRecognizerBase {
