+++ src/sdk/SpeechRecognizer.ts
@@ -9,6 +9,7 @@ import {
     RecognizerConfig,
     ServiceRecognizerBase,
     SpeechConnectionFactory,
+    SpeechServiceConfig,
     SpeechServiceRecognizer,
 } from "../common.speech/Exports";
 import { marshalPromiseToCallbacks } from "../common/Exports";
@@ -276,6 +277,10 @@ export class SpeechRecognizer extends Recognizer {
         }
 
         await super.dispose(disposing);
+    }
+
+    protected createRecognizerConfig(speechConfig: SpeechServiceConfig): RecognizerConfig {
+        return new RecognizerConfig(speechConfig, this.privProperties);
     }
 
     protected createServiceRecognizer(
